+++ CHANGELOG.md
@@ -13,6 +13,8 @@ CHANGELOG for Sulu
     * FEATURE     #1606 [SmartContent]    Added filter by categories to SmartContent Component
     * FEATURE     #1606 [CategoryBundle]  Added TwigExtension to handle categories in twig templates
     * BUGFIX      #1654 [ContentBundle]   Added more path replacers
+    * BUGFIX      #1656 [ContentBundle]   Fixed preview nested properties
+    * BUGFIX      #1656 [ContentBundle]   Fixed preview property attributes
     * BUGFIX      #1649 [ContentBundle]   Fixed floating of block type select
     * BUGFIX      #1650 [LocationBundle]  Fixed configure overlay open multiple times 
     * BUGFIX      #1650 [ContentBundle]   Fixed serialization of null values